Refactor: create new package
`config`.
