+++ backend/src/main/java/dev/lampirg/webgraph/config/WebClientProvider.java
@@ -1,4 +1,4 @@
-package dev.lampirg.webgraph;
+package dev.lampirg.webgraph.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
